requirements.txt
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlalchemy==2.0.36
fastapi==0.115.6
uvicorn==0.32.1
slowapi==0.1.9
